fixing commit errors
